Added Jingle related projects. Fixes to compile in Windows.
git-svn-id: http://yate.null.ro/svn/yate/trunk@1083 acf43c95-373e-0410-b603-e72c3f656dc1

+++ contrib/yjingle/jgengine.cpp
@@ -166,7 +166,7 @@ void JGEngine::createSessionId(String& id)
 {
     Lock lock(m_sessionIdMutex);
     id = "JG";
-    id << m_sessionId << "_" << (int)random();
+    id << (unsigned int)m_sessionId << "_" << (int)random();
     m_sessionId++;
 }
